refactor: move card content into separate files
