+++ src/components/optimizerTab/OptimizerTab.jsx
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types'
 import { Flex } from 'antd'
 import 'ag-grid-community/styles/ag-grid.css'
 
-import OptimizerForm from 'components/OptimizerForm'
+import OptimizerForm from 'components/optimizerTab/OptimizerForm'
 import OptimizerBuildPreview from 'components/OptimizerBuildPreview'
 import Sidebar from 'components/optimizerTab/Sidebar'
 import { OptimizerGrid } from 'components/optimizerTab/OptimizerGrid'
